Summary (required)
First version. The two complex numbers c & z are appended together, to be a 4 dimensional vector.
It always shows you a crosssection of that 4 dimensional fractal.
To switch through those crosssections, use the keys R & F.
Use W, A, S & D to move around the current crosssection, and use the arrow keys to move across the other 2 dimensions.
You can also zoom in and out with Q & E.

Known issues:
- Power hog
- Resolution problems when zooming in close due to minor floating point accuracy inaccuracies