Fixed the shader
idk but it didn't like my struct instance declaration. I previously worked on my laptop, which had an nVidia GPU, so it might be that. I also adjusted the speed (cause my pc faster than my laptop (duh)) and added the pycache to the gitignore cause I hate having to uncheck the pycache